requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-ckeditor==6.1.0
django-redis==5.0.0
googletrans==3.1.0a0
gunicorn==20.1.0

